Move computing basELAyer and MatchExpression functions to utils
